+++ src/ai-dialog-body.js
@@ -1,0 +1,11 @@
+import {customElement, inlineView} from 'aurelia-templating';
+
+@customElement('ai-dialog-body')
+@inlineView(`
+  <template>
+    <slot></slot>
+  </template>
+`)
+export class AiDialogBody {
+
+}