+++ src/aurelia-dialog.js
@@ -1,21 +1,23 @@
 import {DialogConfiguration} from './dialog-configuration';
-export {AiDialog} from './resources/ai-dialog';
-export {AiDialogHeader} from './resources/ai-dialog-header';
-export {AiDialogBody} from './resources/ai-dialog-body';
-export {AiDialogFooter} from './resources/ai-dialog-footer';
-export {AttachFocus} from './resources/attach-focus';
+export {AiDialog} from './ai-dialog';
+export {AiDialogHeader} from './ai-dialog-header';
+export {AiDialogBody} from './ai-dialog-body';
+export {AiDialogFooter} from './ai-dialog-footer';
+export {AttachFocus} from './attach-focus';
 
 export function configure(aurelia, callback) {
   let config = new DialogConfiguration(aurelia);
 
   if (typeof callback === 'function') {
     callback(config);
-    return;
+  } else {
+    config.useDefaults();
   }
 
-  config.useDefaults();
+  config._apply();
 }
 
 export {DialogConfiguration} from './dialog-configuration';
 export {DialogService} from './dialog-service';
-export {DialogController, DialogResult} from './dialog-controller';
+export {DialogController} from './dialog-controller';
+export {DialogResult} from './dialog-result';