+++ src/dialog-service.js
@@ -2,17 +2,26 @@ import {Origin} from 'aurelia-metadata';
 import {Container} from 'aurelia-dependency-injection';
 import {CompositionEngine, ViewSlot} from 'aurelia-templating';
 import {DialogController} from './dialog-controller';
-import {Renderer} from './renderers/renderer';
+import {Renderer} from './renderer';
 import {invokeLifecycle} from './lifecycle';
+import {DialogResult} from './dialog-result';
 
 /**
  * A service allowing for the creation of dialogs.
- * @constructor
  */
 export class DialogService {
   static inject = [Container, CompositionEngine];
 
-  constructor(container: Container, compositionEngine) {
+  /**
+   * The current dialog controllers
+   */
+  controllers: DialogController[];
+  /**
+   * Is there an active dialog
+   */
+  hasActiveDialog: boolean;
+
+  constructor(container: Container, compositionEngine: CompositionEngine) {
     this.container = container;
     this.compositionEngine = compositionEngine;
     this.controllers = [];
@@ -24,20 +33,23 @@ export class DialogService {
    * @param settings Dialog settings for this dialog instance.
    * @return Promise A promise that settles when the dialog is closed.
    */
-  open(settings?: Object) {
+  open(settings?: Object): Promise<DialogResult> {
     let dialogController;
 
     let promise = new Promise((resolve, reject) => {
       let childContainer = this.container.createChild();
       dialogController = new DialogController(childContainer.get(Renderer), settings, resolve, reject);
       childContainer.registerInstance(DialogController, dialogController);
+      let host = dialogController.renderer.getDialogContainer();
 
       let instruction = {
         container: this.container,
         childContainer: childContainer,
         model: dialogController.settings.model,
+        view: dialogController.settings.view,
         viewModel: dialogController.settings.viewModel,
-        viewSlot: new ViewSlot(dialogController._renderer.getDialogContainer(), true)
+        viewSlot: new ViewSlot(host, true),
+        host: host
       };
 
       return _getViewModel(instruction, this.compositionEngine).then(returnedInstruction => {
@@ -53,7 +65,7 @@ export class DialogService {
               dialogController.controller = controller;
               dialogController.view = controller.view;
 
-              return dialogController._renderer.showDialog(dialogController);
+              return dialogController.renderer.showDialog(dialogController);
             });
           }
         });