+++ src/resources/ai-dialog-body.js
@@ -1,6 +1,0 @@
-import {customElement} from 'aurelia-templating';
-
-@customElement('ai-dialog-body')
-export class AiDialogBody {
-
-}