+++ src/resources/ai-dialog.js
@@ -1,6 +1,0 @@
-import {customElement} from 'aurelia-templating';
-
-@customElement('ai-dialog')
-export class AiDialog {
-
-}