+++ test/unit/dialog-configuration.spec.js
@@ -1,10 +1,8 @@
 import {DialogConfiguration} from '../../src/dialog-configuration';
 import {dialogOptions} from '../../src/dialog-options';
-import {DialogRenderer} from '../../src/renderers/dialog-renderer';
-import {initialize} from 'aurelia-pal-browser';
-import {Renderer} from '../../src/renderers/renderer';
-
-initialize();
+import {DialogRenderer} from '../../src/dialog-renderer';
+import {Renderer} from '../../src/renderer';
+import {DOM} from 'aurelia-pal';
 
 let defaultDialogOptions = Object.assign({}, dialogOptions);
 
@@ -16,8 +14,6 @@ let aurelia = {
 describe('DialogConfiguration', () => {
   let configuration;
   beforeEach(() => {
-    initialize();
-
     configuration = new DialogConfiguration(aurelia);
 
     Object.keys(dialogOptions).forEach((key) => {
@@ -32,6 +28,7 @@ describe('DialogConfiguration', () => {
       let renderer = {};
       spyOn(aurelia, 'singleton');
       configuration.useRenderer(renderer);
+      configuration._apply();
       expect(aurelia.singleton).toHaveBeenCalledWith(Renderer, renderer);
     });
 
@@ -39,6 +36,7 @@ describe('DialogConfiguration', () => {
       let renderer = {};
       let settings = { first: 'first', second: 'second' };
       configuration.useRenderer(renderer, settings);
+      configuration._apply();
       expect(dialogOptions.first).toBe(settings.first);
       expect(dialogOptions.second).toBe(settings.second);
     });
@@ -48,6 +46,7 @@ describe('DialogConfiguration', () => {
     it('should call globalResources', () => {
       spyOn(aurelia, 'globalResources');
       configuration.useResource('ai-dialog');
+      configuration._apply();
       expect(aurelia.globalResources).toHaveBeenCalled();
     });
   });
@@ -58,12 +57,42 @@ describe('DialogConfiguration', () => {
       spyOn(configuration, 'useResource').and.callThrough();
 
       configuration.useDefaults();
+      configuration._apply();
       expect(configuration.useRenderer).toHaveBeenCalledWith(DialogRenderer);
       expect(configuration.useResource).toHaveBeenCalledWith('ai-dialog');
       expect(configuration.useResource).toHaveBeenCalledWith('ai-dialog-header');
       expect(configuration.useResource).toHaveBeenCalledWith('ai-dialog-footer');
       expect(configuration.useResource).toHaveBeenCalledWith('ai-dialog-body');
       expect(configuration.useResource).toHaveBeenCalledWith('attach-focus');
+    });
+    
+    it('should inject default style', () => {
+      spyOn(DOM, 'injectStyles').and.callThrough();
+      
+      configuration.useDefaults();
+      configuration._apply();
+      expect(DOM.injectStyles.calls.any()).toEqual(true);
+    });
+  });
+  
+  describe('useCSS', () => {
+    it('should skip injecting empty style', () => {
+      spyOn(DOM, 'injectStyles').and.callThrough();
+      
+      // Undefined css
+      configuration.useCSS(undefined);
+      configuration._apply();
+      expect(DOM.injectStyles.calls.any()).toEqual(false);
+      
+      // Null css
+      configuration.useCSS(null);
+      configuration._apply();
+      expect(DOM.injectStyles.calls.any()).toEqual(false);
+      
+      // Empty string
+      configuration.useCSS('');
+      configuration._apply();
+      expect(DOM.injectStyles.calls.any()).toEqual(false);
     });
   });
 });