+++ test/unit/dialog-renderer.spec.js
@@ -1,24 +1,24 @@
 import {DialogController} from '../../src/dialog-controller';
-import {DialogRenderer} from '../../src/renderers/dialog-renderer';
+import {DialogRenderer} from '../../src/dialog-renderer';
 import {dialogOptions} from '../../src/dialog-options';
 import {configure} from '../../src/aurelia-dialog';
-import {initialize} from 'aurelia-pal-browser';
 import {TestElement} from '../fixtures/test-element';
-
-initialize();
 
 let defaultSettings = {
   lock: true,
   centerHorizontalOnly: false,
-  startingZIndex: 1000
+  startingZIndex: 1000,
+  ignoreTransitions: false
 };
 let newSettings = {
   lock: false,
   centerHorizontalOnly: true,
-  startingZIndex: 1
+  startingZIndex: 1,
+  ignoreTransitions: false
 };
 
 let frameworkConfig = {
+  singleton: () => {},
   globalResources: () => {},
   container: {
     registerInstance: (Type, callback) => {},
@@ -27,10 +27,6 @@ let frameworkConfig = {
 };
 
 describe('the Dialog Renderer', () => {
-  beforeEach(() => {
-    initialize();
-  });
-
   it('uses the default settings', () => {
     let renderer = new DialogRenderer();
     expect(renderer.defaultSettings).toEqual(defaultSettings);
@@ -50,7 +46,7 @@ describe('the Dialog Renderer', () => {
     let controller = new DialogController(renderer, settings, Function.prototype, Function.prototype);
     controller.slot = {
       attached: Function.prototype,
-      anchor: document.createElement('ai-dialog-container')
+      anchor: document.createElement('ai-dialog')
     };
 
     renderer.showDialog(controller);